Cleaned up the description fo C-c C-t for changing TODO states.
There was a misunderstanding in Bastiens fix: C-c C-t is already
defined earlier in that same table.  Using `C-c C-t' is really
discussed later, in the section about todo selection.  There it is
also mentioned that you can use the variable
`org-fast-tag-selection-include-todo' to combine todo states and tags
into a single interface.