Update `org-agenda-to-appt'.
Added `refresh' as the first argument for this function.
A value of `t' for the filter argument now means: prompt
the user for a regular expression and filter out entries
that don't match it.

Don't run `appt-check' inside `org-agenda-to-appt' since
appt-check might already find agenda appointments in the
diary file if it contains (org-diary ...).

Removed the declare-function for appt-check.

Cleaned up the ChangeLog for these changes.