Bug fix: don't trow an error while trying to complete keywords.
When using `org-complete' on keywords like #+TBLFM, a second call to
`org-complete' would throw an error with no helpful message.  It now
fails silently.

I also moved #+DATE to the list of options, since the user might expect
that using `org-complete' twice on it will insert the default date.

Other changes: removed (current-time) from (format-time-string ...)
sexp since it is not necessary.  As far as I checked, there is no
compatibility issue with this.