Doc fixes by Philip Rooke <phil@yax.org.uk>
A number of small changes.  Mainly they are to pacify
'checkdoc-ispell' (if it seems appropriate) and to standardise the
names of the packages you can link to from within Org-mode.  I am
using the package names as spelt or abbreviated in their own
documentation (if I have been able to find it easily).  Thus, for
example, prefer MH-E over MHE, Rmail over RMAIL etc.