Bug fixes for org-publish.el.
`org-publish-get-base-files-1': get the proper list of files when
recursing thru a directory.

`org-publish-get-base-files: use the :exclude property to skip both
files and directories.  This might be refined when the semantic of
publishing options gets simpler, but do this for now.

`org-publish-project': removed third arg of defalias to avoid
compatibility issue for Emacs 22.1.