Implement org-info.js support.  And bug fixes.
Now you can set this stuff directly from the org-mode file.
Documentation is still missing, this is a work-in-progress
commit and still needs check with Sebastian's script.

Bug fix with note taking when auto-repeating a task:
There was a problem if the state change already triggered the
recording of a time, then the request for a *note* in
`org-log-repeat' was ignored.  Reported by Bernt Hansen.