Fix bug with tags filtering.
So far, we had mad invisible the line including the \n after it.
This is a bad idea.  When the cursor is at the beginning of the
line that is preceded by an invisible line, the local text
properties are actually those of the invisible line
just before the current.

Now we make invisible the line together with the \n *before* it,
this should fix things all right.