New default for `org-file-apps'.
The new default asks to open all those files in Emacs, for which
`auto-mode-alist' does contain an entry.  The reasoning is that
Org-mode users are Emacs users and probably like to use exactly
this setup.  The only exception configured as such by default is
that files with html or xhtml extensions will be opened using
the system default, most likely a browser.