Don't keep buffers visited during index publishing.
The index generation of org-publish does briefly visit all files in
order to extract the correct page title.  Visiting is necessary,
because the title may be set by the first line in the buffer, by a
Before this patch, this would cause the generation of a large number
of buffers, one for each file mentioned in the index.  This patch
arranges for these buffers to be removed again.  However, buffers
which were already present will not be removed.

There is still one open problem:  The files to be published are
visited twice, one for publishing them, a second time for creating the
index.  Visiting causes some overhead, and we would like to limit this
overhead.  For now, however, this is not done.