Abbreviate absolute file names in links.
This is a bug fix.  When the option `org-link-file-path-type' has the
value `adaptive', absolute file names would not be abbreviated if they
are relative to the users home directory.  For any other values of
this variable except `noabbrev', absolute links are in fact
abbreviated, so the previous behavior is an inconsistency,
as noticed by Matt Lundin.

This patch fixes this problem and makes sure that also in this case,
file names are abbreviated.

Finally, the patch also fixes a structural bug that would ignore
the double C-u prefix if it was given.