Show state changes in agenda logbook mode.
Org records (depending on setup) various information about progress
achieved with projects and tasks.  It can record when an item is done
(when it is "closed"), it can record time spent working on these
entries (clocking), and it can record state changes.

In the daily/weekly agenda, turning on logbook mode with the `l' key
will lists the tasks closed or clocked on a day.

This patch implements showing recorded state changes in the agenda
logbook.  By default, showing state changes is not turned on.  But you
can configure the new variable `org-agenda-log-mode-items' to trigger
inclusion of state changes.  An alternative way is to press `C-u l' in
the agenda, to temporarily force inclusion of all possible entries
(closed, clock, state) into the agenda display.

In passing, this patch renames the function `org-agenda-get-closed' to
the now more appropriate `org-agenda-get-progress'.  The old name
continues to be available as an alias, but its use is deprecated.