Allow links to contain brackets
Links in the standard form [[link][description]] are allowed to have
brackets in them when these are escapes %5B and %5D.  Creating links
through the Org link interface automatically does this conversion.

However, as noticed by Paul R, there is a bug when actually following
such links, when the link will be cut off at the first closing bracket
in the link.

This patch fixes this problem by adjusting the appropriate regexp.