Coderefs: Implement line numbering and code references
This patch covers the much of the new line numbering and code line
referencing.

We introduce a new function with the longish name
`org-export-replace-src-segments-and-examples'.  It is an extension of
the earlier `org-export-replace-src-segments'.  It now also covers
EXAMPLE blocks, not only source code blocks.  This was necessary to
allow line numbering also in EXAMPLE blocks.

Both clock types now accept a string with options which will be given
in the BEGIN line.  Options currently recognized are -n, +n, and -r.

A bit unsatisfactory about this patch may be that LaTeX-specific code
is now defined in org-exp.el, maybe eventually we want to think of a
way to move this code back to org-export-latex.el.  Really, all it
does is adding line numbers and wrapping them into a verbatim
environment.

Besides adding line numbers in a backend specific way, this patch also
adds code to search or special cookies in source code, like "((1))" or
"((name))".  These are labels that can be addressed by links as
references to specific line in source code.