Coderefs: Supporting functionality
This patch adds support functionality for the new code references.  It
makes sure that `C-c C-o' will follow a code reference correctly.  It
also extends the command org-store-link to actually create a new code
reference cookie when used in a source code temporary buffer created
with "C-c '".  The new link will be stored, so that it can be inserted
with `C-c C-l'.  When inserting such a link, no description will be
prompted for, because these coderef links should best not have a
description part.