Footnotes: Allow validated XHTML with multiple references to one note
Org-mode allows multiple references to the same footnote.  However,
the HTML exporter cannot use the same name for all these references,
because names have to be unique.  This patch appends numbers to
footnote reference names, to make sure they remain unique.

The dark side of this patch means that from the footnote definition,
there is only a single link, to the first reference of the note.
However, in a browser you can always press BACK to go back after
jumping to a footnote definition.