org-remember: Ignore gnus link if not on a message
This allows you to use org-remember in any gnus buffer - even when
not on a location that will create a usable link.  Instead of
failing with the error "Not on a group" just create a remember
note with no link instead.

This makes org-remember more user-friendly - we don't have to
change buffers to a non-gnus buffer to create a remember note.